+++ EnhanceQoL/NewSettingsTable.lua
@@ -1,5 +1,8 @@
 local addonName, addon = ...
 
 addon.variables.NewVersionTableEQOL = {
-
+	EQOL_PROFILES = true,
+	EQOL_ProfilesAddOn = true,
+	EQOL_globalFontFace = true,
+	EQOL_globalFontStyle = true,
 }